feat: 환자 및 진료기록 관리 API

+++ app/models/ai_analysis_result.py
@@ -1,6 +1,6 @@
 from decimal import Decimal
 
-from sqlalchemy import BigInteger, Boolean, DECIMAL, ForeignKey, String
+from sqlalchemy import Boolean, DECIMAL, ForeignKey, Integer, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.core.db.databases import Base
@@ -10,9 +10,9 @@ from app.core.db.models import TimestampMixin
 class AiAnalysisResult(TimestampMixin, Base):
     __tablename__ = "ai_analysis_results"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     record_id: Mapped[int] = mapped_column(
-        BigInteger, ForeignKey("medical_records.id"), nullable=False
+        Integer, ForeignKey("medical_records.id"), nullable=False
     )
     is_pneumonia: Mapped[bool] = mapped_column(Boolean, nullable=False)
     confidence: Mapped[Decimal] = mapped_column(DECIMAL(5, 2), nullable=False)
