+++ app/models/patient.py
@@ -1,4 +1,4 @@
-from sqlalchemy import BigInteger, SmallInteger, String
+from sqlalchemy import Integer, SmallInteger, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.core.db.databases import Base
@@ -9,7 +9,7 @@ from app.models.user import Gender
 class Patient(TimestampMixin, Base):
     __tablename__ = "patients"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     name: Mapped[str] = mapped_column(String(30), nullable=False)
     age: Mapped[int] = mapped_column(SmallInteger, nullable=False)
     gender: Mapped[Gender] = mapped_column(nullable=False)
